+++ client/dns.go
@@ -3,10 +3,7 @@ package client
 import (
 	"fmt"
 	"log"
-	"math"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/go-routeros/routeros"
 )
@@ -16,39 +13,6 @@ type DnsRecord struct {
 	Name    string
 	Ttl     int `mikrotik:"ttl,ttlToSeconds"`
 	Address string
-}
-
-func ttlToSeconds(ttl string) int {
-	parts := strings.Split(ttl, "d")
-
-	idx := 0
-	days := 0
-	var err error
-	if len(parts) == 2 {
-		idx = 1
-		days, err = strconv.Atoi(parts[0])
-
-		// We should be parsing an ascii number
-		// if this fails we should fail loudly
-		if err != nil {
-			panic(err)
-		}
-
-		// In the event we just get days parts[1] will be an
-		// empty string. Just coerce that into 0 seconds.
-		if parts[1] == "" {
-			parts[1] = "0s"
-		}
-	}
-	d, err := time.ParseDuration(parts[idx])
-
-	// We should never receive a duration greater than
-	// 23h59m59s. So this should always parse.
-	if err != nil {
-		panic(err)
-	}
-	return 86400*days + int(d)/int(math.Pow10(9))
-
 }
 
 func (client Mikrotik) AddDnsRecord(name, address string, ttl int) (*routeros.Reply, error) {
